+++ cicd_policy_gate/dbt_project/models/marketing/active_customer_activation.sql
@@ -1,0 +1,1 @@
+SELECT customer_name, email FROM customers WHERE marketing_consent = 'opted_in'