+++ cicd_policy_gate/dbt_project/models/marts/eu_customer_arr_detail.sql
@@ -1,0 +1,1 @@
+SELECT customer_name, email FROM customers WHERE region = 'EU'